+++ lib/calc/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export { computeMetrics, metricDen, parseMetricValue } from './metrics';
+export { formatINR, formatINRShort } from './currency';